requirements.txt
requests
streamlit
uvicorn
fastapi
torch
torchvision
Pillow
streamlit
streamlit-drawable-canvas